📝 Update Modal showcase
