+++ apps/cookbook/src/app/examples/modal-example/modal-example.component.scss
@@ -1,4 +1,8 @@
-@import './libs/designsystem/src/lib/scss/utils';
+@import '../examples.shared';
+
+h1 {
+  text-align: center;
+}
 
 @include media('>medium') {
   .mobile {
